Update README.md
I added a small change to user input for hot fluid and cold fluid substances. The way it was before was if the user input anything else other than the available substances, they program would crash. Now, if they input anything other than the available substances, it re-prompts them to enter the substances again. Also, I changed the "or" back to an "and" because with "or" if the user inputs either K or F, the program would always re-prompt the user to "Enter Hot Fluid inlet Temp & Units (F or K): " prompt. This happens because the "or" will take whichever side is true, and restart the prompt. If we enter "F", then F is not "K", which is True, and it will print "invalid units." I know "and" seems counter-intuitive, and I don't remember how it works again, but I worked this one out with Logan the TA.